test: add struct size validation for generated types
Add layout_test.go with 32 struct size checks against C ABI on
linux/amd64. Fix two layout bugs found by the tests:
- CEFKeyEventT: character/unmodified_character are char16_t (uint16),
  not pointer-sized
- CEFAcceleratedPaintInfoCommonT: has_* flags are uint8_t, not
  pointer-sized

+++ internal/capi/cef_types.go
@@ -355,8 +355,8 @@ type CEFKeyEventT struct {
 	WindowsKeyCode       int32
 	NativeKeyCode        int32
 	IsSystemKey          int32
-	Character            uintptr
-	UnmodifiedCharacter  uintptr
+	Character            uint16
+	UnmodifiedCharacter  uint16
 	FocusOnEditableField int32
 }
